+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Integration with Ipfs Fetcher
+- Consuming delete proposal events based on snapshot messages
+
 ## [0.5.0-draft] - 2024-03-12
 
 ### Added